[refactor] added ec package to primitives and wif to compat
