+++ compat/bip32/extendedkey.go
@@ -22,7 +22,7 @@ import (
 	"github.com/bitcoin-sv/go-sdk/base58"
 	"github.com/bitcoin-sv/go-sdk/chaincfg"
 	"github.com/bitcoin-sv/go-sdk/crypto"
-	"github.com/bitcoin-sv/go-sdk/ec"
+	ec "github.com/bitcoin-sv/go-sdk/primitives/ec"
 )
 
 const (
