+++ docs/examples/create_tx_with_op_return/create_tx_with_op_return.go
@@ -4,7 +4,7 @@ import (
 	"context"
 	"log"
 
-	"github.com/bitcoin-sv/go-sdk/ec/wif"
+	wif "github.com/bitcoin-sv/go-sdk/compat/wif"
 	"github.com/bitcoin-sv/go-sdk/transaction"
 	"github.com/bitcoin-sv/go-sdk/transaction/unlocker"
 )
