+++ primitives/ec/privatekey_test.go
@@ -1,4 +1,4 @@
-package ec
+package primitives
 
 import (
 	"bytes"
