+++ primitives/ec/secp256k1.go
@@ -1,4 +1,4 @@
-package ec
+package primitives
 // Copyright (c) 2015 The btcsuite developers
 // Use of this source code is governed by an ISC
 // license that can be found in the LICENSE file.
