+++ primitives/ec/testdata/BRC42.private.vectors.json
