+++ primitives/ec/testdata/BRC42.public.vectors.json
