+++ primitives/ec/testdata/SymmetricKey.vectors.json
